fix for and while workgraph
